minor: improvements to release notes specific tests
