+++ releasenotes-builder/src/test/java/com/github/checkstyle/internal/AbstractReleaseNotesTestSupport.java
@@ -52,6 +52,7 @@ import org.mockito.stubbing.Answer;
 import com.github.checkstyle.Main;
 import com.github.checkstyle.git.CsGit;
 import com.github.checkstyle.github.CsGitHub;
+import com.github.checkstyle.globals.Constants;
 import com.github.checkstyle.publishers.TwitterPublisher;
 
 public abstract class AbstractReleaseNotesTestSupport extends AbstractPathTestSupport {
@@ -88,6 +89,15 @@ public abstract class AbstractReleaseNotesTestSupport extends AbstractPathTestSu
         + " -validateVersion                  Whether release number should be%n"
         + "                                   validated to match release notes%n"
         + "                                   labels.%n");
+
+    protected static final String BUG = Constants.BUG_LABEL;
+    protected static final String NEW_FEATURE = Constants.NEW_FEATURE_LABEL;
+    protected static final String NEW_MODULE = Constants.NEW_MODULE_LABEL;
+    protected static final String MISC = Constants.MISCELLANEOUS_LABEL;
+    protected static final String BREAKING = Constants.BREAKING_COMPATIBILITY_LABEL;
+
+    protected static final String MSG_EXECUTION_SUCCEEDED = System.lineSeparator()
+            + "Execution succeeded!" + System.lineSeparator();
 
     private static final Set<RevCommit> TEST_COMMITS = new LinkedHashSet<>();
 
@@ -232,4 +242,31 @@ public abstract class AbstractReleaseNotesTestSupport extends AbstractPathTestSu
         );
         runMainAndAssertReturnCode(expectedExitCode, listOfArgs.toArray(new String[0]));
     }
+
+    protected static String getExecutionFailedMessage(int amountOfErrors) {
+        return System.lineSeparator()
+            + "Generation ends with " + amountOfErrors + " errors."
+            + System.lineSeparator();
+    }
+
+    protected static String getReleaseIsMinorMessage(String releaseVersion) {
+        return System.lineSeparator()
+            + "[ERROR] Validation of release number failed. "
+            + "Release number is minor(" + releaseVersion + "), but release notes do not contain "
+            + "'new' or 'breaking compatability' labels. Please correct release number by running "
+            + "https://github.com/checkstyle/checkstyle/actions/workflows/bump-version-and-"
+            + "update-milestone.yml"
+            + System.lineSeparator();
+    }
+
+    protected static String getReleaseIsPatchMessage(String releaseVersion) {
+        return System.lineSeparator()
+            + "[ERROR] Validation of release number failed. "
+            + "Release number is a patch(" + releaseVersion + "), but release notes contain 'new' "
+            + "or 'breaking compatability' labels. Please correct release number by running "
+            + "https://github.com/checkstyle/checkstyle/actions/workflows/bump-version-and-"
+            + "update-milestone.yml"
+            + System.lineSeparator();
+    }
+
 }
